rename sim_phenotypes module to sim_phenotype (singular)
